Struct darkfi_contract_test_harness::TestHarness

pub struct TestHarness {
    pub holders: HashMap<Holder, Wallet>,
    pub proving_keys: HashMap<String, (ProvingKey, ZkBinary)>,
    pub genesis_block: BlockInfo,
    pub verify_fees: bool,
}

Native contract test harness instance

Fields
holders: HashMap<Holder, Wallet>

Initialized Holders for this instance

proving_keys: HashMap<String, (ProvingKey, ZkBinary)>

Cached [ProvingKey]s for native contract ZK proving

genesis_block: BlockInfo

The genesis block for this harness

verify_fees: bool

Marker to know if we’re supposed to include tx fees

Implementations

impl TestHarness

async fn pow_reward( &mut self, holder: &Holder, recipient: Option<&Holder>, reward: Option<u64>, fees: Option<u64>, ) -> Result<(Transaction, MoneyPoWRewardParamsV1)>

Create a Money::PoWReward transaction for a given Holder.

Optionally takes a specific reward recipient and a nonstandard reward value. Returns the created [Transaction] and [MoneyPoWRewardParamsV1].

pub async fn generate_block( &mut self, miner: &Holder, holders: &[Holder], ) -> Result<Vec<OwnCoin>>

Generate and add an empty block to the given Holders blockchains. The miner holder will produce the block and receive the reward.

Returns any found [OwnCoin]s.

impl TestHarness

pub async fn create_empty_fee_call( &mut self, holder: &Holder, ) -> Result<(Transaction, MoneyFeeParamsV1)>

Create an empty transaction that includes a Money::Fee call. This is generally used to test the actual fee call, and also to see the gas usage of the call without other parts.

pub async fn execute_empty_fee_call_tx( &mut self, holder: &Holder, tx: Transaction, params: &MoneyFeeParamsV1, block_height: u32, ) -> Result<Vec<OwnCoin>>

Execute the transaction created by create_empty_fee_call() for a given Holder

Returns any found [OwnCoin]s.

pub async fn append_fee_call( &mut self, holder: &Holder, tx: Transaction, block_height: u32, spent_coins: &[OwnCoin], ) -> Result<(ContractCall, Vec<Proof>, Vec<SecretKey>, Vec<OwnCoin>, MoneyFeeParamsV1)>

Create and append a Money::Fee call to a given [Transaction] for a given Holder.

Additionally takes a set of spent coins in order not to reuse them here.

Returns the Fee call, and all necessary data and parameters related.

impl TestHarness

pub async fn genesis_mint( &mut self, holder: &Holder, amount: u64, spend_hook: Option<FuncId>, user_data: Option<Base>, ) -> Result<(Transaction, MoneyGenesisMintParamsV1)>

Create a Money::GenesisMint transaction for a given Holder.

Returns the created [Transaction] and its parameters.

pub async fn execute_genesis_mint_tx( &mut self, holder: &Holder, tx: Transaction, params: &MoneyGenesisMintParamsV1, block_height: u32, append: bool, ) -> Result<Vec<OwnCoin>>

Execute the [Transaction] created by genesis_mint().

Returns any found [OwnCoin]s.

impl TestHarness

pub async fn transfer( &mut self, amount: u64, holder: &Holder, recipient: &Holder, owncoins: &[OwnCoin], token_id: TokenId, block_height: u32, half_split: bool, ) -> Result<(Transaction, (MoneyTransferParamsV1, Option<MoneyFeeParamsV1>), Vec<OwnCoin>)>

Create a Money::Transfer transaction.

pub async fn execute_transfer_tx( &mut self, holder: &Holder, tx: Transaction, call_params: &MoneyTransferParamsV1, fee_params: &Option<MoneyFeeParamsV1>, block_height: u32, append: bool, ) -> Result<Vec<OwnCoin>>

Execute a Money::Transfer transaction for a given Holder.

Returns any found [OwnCoin]s.

impl TestHarness

pub async fn token_mint( &mut self, amount: u64, holder: &Holder, recipient: &Holder, token_blind: BaseBlind, spend_hook: Option<FuncId>, user_data: Option<Base>, block_height: u32, ) -> Result<(Transaction, MoneyTokenMintParamsV1, MoneyAuthTokenMintParamsV1, Option<MoneyFeeParamsV1>)>

Mint an arbitrary token for a given recipient using Money::TokenMint

pub async fn execute_token_mint_tx( &mut self, holder: &Holder, tx: Transaction, mint_params: &MoneyTokenMintParamsV1, auth_params: &MoneyAuthTokenMintParamsV1, fee_params: &Option<MoneyFeeParamsV1>, block_height: u32, append: bool, ) -> Result<Vec<OwnCoin>>

Execute the transaction created by token_mint() for a given Holder.

Returns any found [OwnCoin]s.

pub async fn token_freeze( &mut self, holder: &Holder, block_height: u32, ) -> Result<(Transaction, MoneyAuthTokenFreezeParamsV1, Option<MoneyFeeParamsV1>)>

Freeze the supply of a minted token

pub async fn execute_token_freeze_tx( &mut self, holder: &Holder, tx: Transaction, _freeze_params: &MoneyAuthTokenFreezeParamsV1, fee_params: &Option<MoneyFeeParamsV1>, block_height: u32, append: bool, ) -> Result<Vec<OwnCoin>>

Execute the transaction created by token_freeze() for a given Holder.

Returns any found [OwnCoin]s.

impl TestHarness

pub async fn otc_swap( &mut self, holder0: &Holder, owncoin0: &OwnCoin, holder1: &Holder, owncoin1: &OwnCoin, block_height: u32, ) -> Result<(Transaction, MoneyTransferParamsV1, Option<MoneyFeeParamsV1>)>

Create a Money::OtcSwap transaction with two given Holders.

Returns the [Transaction], and the transaction parameters.

pub async fn execute_otc_swap_tx( &mut self, holder: &Holder, tx: Transaction, swap_params: &MoneyTransferParamsV1, fee_params: &Option<MoneyFeeParamsV1>, block_height: u32, append: bool, ) -> Result<Vec<OwnCoin>>

Execute the transaction created by otc_swap() for a given Holder.

Returns any found [OwnCoin]s.

impl TestHarness

pub async fn deploy_contract( &mut self, holder: &Holder, wasm_bincode: Vec<u8>, block_height: u32, ) -> Result<(Transaction, DeployParamsV1, Option<MoneyFeeParamsV1>)>

Create a Deployooor::Deploy transaction with the given WASM bincode.

Returns the [Transaction], and necessary parameters.

pub async fn execute_deploy_tx( &mut self, holder: &Holder, tx: Transaction, _params: &DeployParamsV1, fee_params: &Option<MoneyFeeParamsV1>, block_height: u32, append: bool, ) -> Result<Vec<OwnCoin>>

Execute the transaction created by deploy_contract() for a given Holder.

Returns any found [OwnCoin]s.

impl TestHarness

pub async fn dao_mint( &mut self, holder: &Holder, dao_info: &Dao, dao_kp: &Keypair, block_height: u32, ) -> Result<(Transaction, DaoMintParams, Option<MoneyFeeParamsV1>)>

Create a Dao::Mint transaction with the given [Dao] info and a keypair. Takes a Holder for optionally paying the transaction fee.

Returns the [Transaction] and any relevant parameters.

pub async fn execute_dao_mint_tx( &mut self, holder: &Holder, tx: Transaction, params: &DaoMintParams, fee_params: &Option<MoneyFeeParamsV1>, block_height: u32, append: bool, ) -> Result<Vec<OwnCoin>>

Execute the transaction created by dao_mint() for a given Holder.

Returns any found [OwnCoin]s.

impl TestHarness

pub async fn dao_propose( &mut self, proposer: &Holder, proposal_coinattrs: &[CoinAttributes], user_data: Base, dao: &Dao, dao_bulla: &DaoBulla, block_height: u32, ) -> Result<(Transaction, (DaoProposeParams, Option<MoneyFeeParamsV1>), DaoProposal)>

Create a Dao::Propose transaction.

pub async fn execute_dao_propose_tx( &mut self, holder: &Holder, tx: Transaction, params: &DaoProposeParams, fee_params: &Option<MoneyFeeParamsV1>, block_height: u32, append: bool, ) -> Result<Vec<OwnCoin>>

Execute the transaction created by dao_propose() for a given Holder.

Returns any found [OwnCoin]s.

impl TestHarness

pub async fn dao_vote( &mut self, voter: &Holder, vote_option: bool, dao: &Dao, dao_keypair: &Keypair, proposal: &DaoProposal, proposal_bulla: &DaoProposalBulla, block_height: u32, ) -> Result<(Transaction, DaoVoteParams, Option<MoneyFeeParamsV1>)>

Create a Dao::Vote transaction.

pub async fn execute_dao_vote_tx( &mut self, holder: &Holder, tx: Transaction, _params: &DaoVoteParams, fee_params: &Option<MoneyFeeParamsV1>, block_height: u32, append: bool, ) -> Result<Vec<OwnCoin>>

Execute the transaction made by dao_vote() for a given Holder.

Returns any found [OwnCoin]s.

impl TestHarness

pub async fn dao_exec( &mut self, holder: &Holder, dao: &Dao, dao_bulla: &DaoBulla, proposal: &DaoProposal, proposal_coinattrs: Vec<CoinAttributes>, yes_vote_value: u64, all_vote_value: u64, yes_vote_blind: ScalarBlind, all_vote_blind: ScalarBlind, block_height: u32, ) -> Result<(Transaction, MoneyTransferParamsV1, DaoExecParams, Option<MoneyFeeParamsV1>)>

Create a Dao::Exec transaction.

pub async fn execute_dao_exec_tx( &mut self, holder: &Holder, tx: Transaction, xfer_params: &MoneyTransferParamsV1, _exec_params: &DaoExecParams, fee_params: &Option<MoneyFeeParamsV1>, block_height: u32, append: bool, ) -> Result<Vec<OwnCoin>>

Execute the transaction made by dao_exec() for a given Holder.

Returns any found [OwnCoin]s.

impl TestHarness

pub async fn new(holders: &[Holder], verify_fees: bool) -> Result<Self>

Instantiate a new TestHarness given a slice of Holders. Additionally, a verify_fees boolean will enforce tx fee verification.

pub fn assert_trees(&self, holders: &[Holder])

Assert that all holders’ trees are the same